stm_copro: Make coprocessor driver work with udev
This patch updates the coprocessor device driver to allow
the dynamic device node creation under /dev.

See bugzilla 3398.

+++ drivers/stm/copro-st_socs.c
@@ -79,7 +79,8 @@ int coproc_cpu_grant(coproc_t * cop, unsigned long arg)
 		bootAddr = arg;
 	/* Now set the less meaningful bit to trigger the ST231 start */
 	bootAddr |= 1;
-	DPRINTK(">>> platform: st231.%u start from 0x%x...\n",id, bootAddr);
+	DPRINTK(">>> platform: st231.%u start from 0x%x...\n",
+					id, (unsigned int)bootAddr);
 	/* bypass the st40 to reset only the coprocessor */
 	sysconf_write(copro_reset_out, 1);
